% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matdat}
\alias{matdat}
\title{Matching Fall 2024}
\format{
\subsection{\code{matdat}}{

A dataframe
}
}
\usage{
matdat
}
\description{
Fall 2024 matching data. Genetic matching. Exact matching on grade and gender.
}
\keyword{datasets}
