% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matdat_exact_loc_r1}
\alias{matdat_exact_loc_r1}
\title{Matching Data Fall 2024}
\format{
\subsection{\code{matdat_exact_loc_r1}}{

A dataframe
}
}
\usage{
matdat_exact_loc_r1
}
\description{
Matching Data Fall 2024
}
\keyword{datasets}
