% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matdat_loc3}
\alias{matdat_loc3}
\title{Matching Data Fall 2024}
\format{
\subsection{\code{matdat_loc3}}{

A dataframe
}
}
\usage{
matdat_loc3
}
\description{
Matching Data Fall 2024
}
\keyword{datasets}
