% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matdat_loc_r1_v2}
\alias{matdat_loc_r1_v2}
\title{Matching Data Fall 2024}
\format{
\subsection{\code{matdat_loc_r1_v2}}{

A dataframe
}
}
\usage{
matdat_loc_r1_v2
}
\description{
Matching Data Fall 2024
}
\keyword{datasets}
