% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matdat_mixed}
\alias{matdat_mixed}
\title{Matching Data Fall 2024}
\format{
\subsection{\code{matdat_mixed}}{

A dataframe
}
}
\usage{
matdat_mixed
}
\description{
Uses 'matdat_exact_loc' for grade 2 and 'matdat_loc' for other grades.
}
\keyword{datasets}
