% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mo_map_fall24}
\alias{mo_map_fall24}
\title{Fall 2024 MO MAP Data}
\format{
\subsection{\code{mo_map_fall24}}{

A dataframe
}
}
\usage{
mo_map_fall24
}
\description{
Fall 2024 MO MAP Data
}
\keyword{datasets}
