% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment_district}
\alias{treatment_district}
\title{Treatment District}
\format{
\subsection{\code{treatment_district}}{

A string with the treatment district ID
}
}
\usage{
treatment_district
}
\description{
Treatment District
}
\keyword{datasets}
